Implement search bar and table view setup logic
