+++ WordPress/Classes/ViewRelated/Me/My Domains/MyDomainsViewController+Strings.swift
@@ -1,0 +1,17 @@
+import Foundation
+
+extension MyDomainsViewController {
+
+    enum Strings {
+        static let title = NSLocalizedString(
+            "domain.management.title",
+            value: "My Domains",
+            comment: "Domain Management Screen Title"
+        )
+        static let searchBar = NSLocalizedString(
+            "domain.management.search-bar.title",
+            value: "Search domains",
+            comment: "The search bar title in My Domains screen."
+        )
+    }
+}
